pyproject.toml
[project]
name = "azure-devops-sprint-mcp"
version = "2.1.0"
description = "Enterprise-grade MCP server for Azure DevOps sprint board and work item management with multi-project support"
authors = [
    { name = "Azure DevOps MCP Contributors", email = "noreply@example.com" }
]
maintainers = [
    { name = "Azure DevOps MCP Contributors", email = "noreply@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["azure-devops", "mcp", "model-context-protocol", "sprint", "work-items", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastmcp>=0.4.0",
    "azure-devops>=7.1.0b4",
    "azure-identity>=1.15.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/azure-devops-sprint-mcp"
Documentation = "https://github.com/yourusername/azure-devops-sprint-mcp/tree/main/docs"
Repository = "https://github.com/yourusername/azure-devops-sprint-mcp"
Issues = "https://github.com/yourusername/azure-devops-sprint-mcp/issues"
Changelog = "https://github.com/yourusername/azure-devops-sprint-mcp/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
